Mostly working
Need to perform code cleanup, and make sure startup of the driver
contains everything it needs to properly verify the status of the
battery.

Could use code to get the voltage etc, however I don't see that code in
the argon40 python stuff so I'd have to investigate the I2C device.