Add support to change the shutdown threshold.
The default value for shutdown is less than 5% of battery life left.
Once the system hits this value (or the one specified by the user) the
system will shutdown.

The soc_shutdown value can be changed to 0 to disable this feature, just
realise that the system will lose power and unsafely shutdown.

Also added install and remove functionality.